pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clipx"
version = "0.1.0"
description = "Image background removal tool supporting U2Net and CascadePSP models"
readme = "README.md"
authors = [
    {name = "zhangsdong", email = "zhangsdong0402@gmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Image Processing",
]
requires-python = ">=3.7"
dependencies = [
    "torch>=1.7.0",
    "torchvision>=0.8.0",
    "pillow>=8.0.0",
    "numpy>=1.19.0",
    "click>=7.1.2",
    "requests>=2.25.0",
    "tqdm>=4.50.0",
]

[project.urls]
"Homepage" = "https://github.com/zhangsdong/clipx"
"Bug Tracker" = "https://github.com/zhangsdong/clipx/issues"
"Source Code" = "https://github.com/zhangsdong/clipx"

[project.scripts]
clipx = "clipx.cli:main"

[tool.setuptools]
packages = ["clipx"]

[tool.setuptools.package-data]
clipx = ["models/**/*.py"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "PIL.*",
    "numpy.*",
    "tqdm.*",
]
ignore_missing_imports = true